requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
zeroconf>=0.131.0
asyncio-dgram>=2.1.2
pydantic>=2.5.0
pyyaml>=6.0.1